Skiller ut typer og useDirection
