+++ src/components/games/SpareslagenSpill/types.ts
@@ -1,0 +1,6 @@
+type Position = {
+  x: number;
+  y: number;
+};
+
+type Direction = "l" | "r" | "u" | "d";
